Tweak int type

+++ src/consensus/params.h
@@ -125,7 +125,7 @@ struct Params {
 
     uint256 nMinimumChainWork;
     uint256 defaultAssumeValid;
-    uint32_t workFinalBlockSubsidy;
+    int64_t workFinalBlockSubsidy;
     bool EnableRBF() const { return fRbfEnabled; }
 };
 } // namespace Consensus
